ActionCompendium is now MetaCompendium, registers defined Actions, Attacks, Spells.
