+++ src/entities/action/weapon_action.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
 from typing import Any, Generator
 
 from src.entities.action.actions import ActionMeta, BaseAction
-from src.entities.combat.attack_types import WeaponAttack
+from src.entities.combat.weapon_attacks import WeaponAttackMeta
 
 Event = dict[str, Any]
 
@@ -24,7 +24,7 @@ class WeaponAttackAction(BaseAction, metaclass=ActionMeta):
 
     @classmethod
     def execute(
-        cls, fighter: Fighter, target: Fighter, attack: WeaponAttack
+        cls, fighter: Fighter, target: Fighter, attack: WeaponAttackMeta
     ) -> Generator[Event]:
         if fighter.action_points.current >= attack.ap_cost:
             fighter.action_points.deduct_cost(cls.cost(fighter))
@@ -34,7 +34,7 @@ class WeaponAttackAction(BaseAction, metaclass=ActionMeta):
             yield {"message": f"Not enough AP for {attack.name}"}
 
     @classmethod
-    def details(cls, fighter: Fighter, attack: WeaponAttack) -> dict:
+    def details(cls, fighter: Fighter, attack: WeaponAttackMeta) -> dict:
         on_confirm = lambda target: fighter.ready_action(cls(fighter, target, attack))
 
         return {
@@ -51,7 +51,9 @@ class WeaponAttackAction(BaseAction, metaclass=ActionMeta):
             for attack in fighter.equipment.weapon.available_attacks
         ]
 
-    def __init__(self, fighter: Fighter, target: Fighter, attack: WeaponAttack) -> None:
+    def __init__(
+        self, fighter: Fighter, target: Fighter, attack: WeaponAttackMeta
+    ) -> None:
         self.fighter = fighter
         self.target = target
         self.attack = attack
