+++ src/tests/fixtures.py
@@ -60,9 +60,7 @@ class EncounterFactory:
             fighter=Fighter(**FighterFixtures.strong(enemy=enemy, boss=False)),
         ).with_inventory_capacity(1)
         weapon = Equippable.init_affixes(None, Sword)
-        e.fighter.equipment = Equipment(
-            e.fighter
-        )
+        e.fighter.equipment = Equipment(e.fighter)
         e.fighter.equipment.equip_item(weapon)
         return e
 
@@ -75,9 +73,7 @@ class EncounterFactory:
             fighter=Fighter(**FighterFixtures.baby(enemy=enemy, boss=False)),
         ).with_inventory_capacity(1)
         weapon = Equippable.init_affixes(None, Sword)
-        e.fighter.equipment = Equipment(
-            e.fighter
-        )
+        e.fighter.equipment = Equipment(e.fighter)
         e.fighter.equipment.equip_item(weapon)
         return e
